% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xes_cases.R
\name{read_xes_cases}
\alias{read_xes_cases}
\title{Case Attributes from Xes-file}
\usage{
read_xes_cases(xesfile = file.choose())
}
\arguments{
\item{xesfile}{Reference to a .xes file, conforming to the xes-standard.}
}
\description{
Extracts case attributes from a xes-file.
}
\seealso{
\url{http://www.xes-standard.org/}
}
