% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xes.R
\name{write_xes}
\alias{write_xes}
\title{Write XES file}
\usage{
write_xes(eventlog, file = file.choose(), case_attributes = NULL)
}
\arguments{
\item{eventlog}{An event log object}

\item{file}{Destination file}

\item{case_attributes}{List of columns containing case_attributes}
}
\description{
Function for writing xes-file
}
