% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xesreadR.R
\docType{package}
\name{xesreadR}
\alias{xesreadR}
\alias{xesreadR-package}
\title{xesreadR - Read and write XES fles}
\description{
Functions for reading and writing XES-files. XES (eXtensible Event Stream) is the IEEE standard for storing and sharing event data (see <http://standards.ieee.org/findstds/standard/1849-2016.html> for more info)
}
