% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xes.R
\name{read_xes}
\alias{read_xes}
\title{read_xes}
\usage{
read_xes(xesfile = file.choose())
}
\arguments{
\item{xesfile}{Reference to a .xes file, conforming to the xes-standard.}
}
\description{
Extracts eventlog from a xes-file.
}
\seealso{
\url{http://www.xes-standard.org/}
}
